+++ client/components/Editor/Pre.tsx
@@ -2,7 +2,10 @@ import React from 'react';
 import { prismSlug } from '../../helpers';
 
 const Pre: React.FC<{ language: string }> = ({ language, children }) => (
-  <pre className={`language-${prismSlug(language)}`} spellCheck={false}>
+  <pre
+    className={`language-${prismSlug(language)} line-numbers`}
+    spellCheck={false}
+  >
     {children}
   </pre>
 );